Update get-started.md
Sorry, for wasting your time, I know I can have mistakes. The point of this pull request was that this tutorial is not as simple as it could be. And this separation between 'you' and 'we' creates strange feeling. I didn't plan to read all this tutorial because I assume everything is written in this manner.