Update build.md
The Button story has no actions defined and the screen grab is actually showing the Link story. It's a bit confusing. Also the excerpted example code for the buttonWrapper story doesn't match up with the codebase. I wasn't sure if the tutorial wanted us to add the action to the Button story or meant to show us the action in the Link story. I took the former approach but the video should be updated