Update screen.md for Angular
There is a "missing" selector that wasn't marked as a change in the docs.  I certainly skipped it so I got an error.

PS: It took me 20 minutes to figure out where to find the source code for the tutorial. A proper link in the web that links here would be appreciated.