Removed styling only addressed in topic 'Addons' later on in the tutorial.
In the topic 'Addons' this component is tested by using the Controls Addon. At this point the overflow behavior with long strings is addressed by adding the styling and a visual regression test. Since this style is already added at the start of the course, someone following along the tutorial will not see the text overflow the component like in the examples. So it should not be there yet at this point.